% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bvec}
\alias{bvec}
\title{typedef \href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<bool> bvec;}
\description{
a vectof of bool's.
}
\seealso{
Other basic-types: \code{\link{uint}}, \code{\link{uvec}},
  \code{\link{vec}}
}
