% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRPackageFromTemplate.R
\name{CreateRPackageFromTemplate}
\alias{CreateRPackageFromTemplate}
\title{Create an R-package based on a template}
\usage{
CreateRPackageFromTemplate(packageName, path = ".",
  class = "MyTraversalTask", overwrite = FALSE)
}
\arguments{
\item{packageName}{a character string denoting the name of the package.}

\item{path}{a character string denoting a directory where to hold the package.}

\item{class}{a character string denoting the name of a TraversalSpecification 
class to be implemented in the package.}
}
\value{
If the function runs correctly nothing. Otherwise an error is raised.
}
\description{
Create an R-package based on a template
}
\details{
This function needs Internet connection to clone the 
\link{https://github.com/venelin/PMMUsingSPLITT.git} repository in \code{path}. 
Then, it replaces the name of the package and the the C++ namespace 
\code{PMMUsingSPLITT} with \code{packageName} and the name of the class 
\code{AbcPMM} with \code{className}. All these arguments must
be valid C++ identifiers. The \code{packageName} parameter must be a valid 
name for an R-package.
}
