% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OrderedTree}
\alias{OrderedTree}
\title{template<class Node, class Length>class OrderedTree: public Tree<Node, Length>}
\description{
A generic C++ class defining the data structure and basic 
  operations with a tree.
}
\section{Template Arguments}{

\describe{
\item{class Node}{see \code{\link{NodeType}}.}
\item{class Length}{see \code{\link{LengthType}}.}
}
}

\section{Public Methods}{

\describe{
}
}

\seealso{
\code{\link{OrderedTree}}
}
