% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotParallelTraversal.R
\name{PlotParallelTraversal}
\alias{PlotParallelTraversal}
\title{Global definitions needed for the R-check to pass}
\usage{
PlotParallelTraversal(tree, x.lim = c(-2, max(node.depth.edgelength(tree)) +
  1), draw_prune_ranges = TRUE)
}
\arguments{
\item{tree}{a phylo object}

\item{x.lim}{limits on the time (x) -axis}

\item{draw_prune_ranges}{logical indicating if prune-ranges should be
denoted on each plot.}
}
\description{
Visualize the parallel traversal of the nodes in the tree from the tips to the root.
}
