% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::bvec}
\alias{SPLITT::bvec}
\title{a vector of \code{bool}s.}
\description{
\code{typedef }\href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<bool> bvec;}
}
\seealso{
\link{SPLITT}

Other basic types: \code{\link{SPLITT::uint}},
  \code{\link{SPLITT::uvec}}, \code{\link{SPLITT::vec}}
}
