% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::EMPTY_UVEC}
\alias{SPLITT::EMPTY_UVEC}
\title{A global constant for the empty \link[=SPLITT::uvec]{\code{uvec}};}
\description{
Currently defined as
\code{const uvec EMPTY_UVEC;}
}
\seealso{
\link{SPLITT}

Other global constants: \code{\link{SPLITT::NA_UINT}}
}
