% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::G_EMPTY_UVEC}
\alias{SPLITT::G_EMPTY_UVEC}
\title{A global constant for the empty \code{\link[=SPLITT::uvec]{uvec}};}
\description{
Currently defined as
\code{const uvec G_EMPTY_UVEC;}
}
\seealso{
\link{SPLITT}

Other global constants: \code{\link{SPLITT::G_NA_UINT}},
  \code{\link{SPLITT::G_PI}}
}
