% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::G_NA_UINT}
\alias{SPLITT::G_NA_UINT}
\title{A global constant for the integer NA.}
\description{
Currently defined as
\code{const uint G_NA_UINT = std::numeric_limits<uint>::max();}
}
\seealso{
\link{SPLITT}

Other global constants: \code{\link{SPLITT::G_EMPTY_UVEC}},
  \code{\link{SPLITT::G_PI}}
}
\concept{global constants}
