% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::G_PI}
\alias{SPLITT::G_PI}
\title{The mathematical constant Pi.}
\description{
Currently defined as
\code{const double G_PI = 3.1415926535897932385;}
}
\seealso{
\code{\link{SPLITT}}

Other global constants: \code{\link{SPLITT::G_EMPTY_UVEC}},
  \code{\link{SPLITT::G_NA_UINT}}
}
