% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::IsNA}
\alias{SPLITT::IsNA}
\title{Check a vector for \code{NA}s.}
\arguments{
\item{x}{a \href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<T> const&} 
of elements to be checked.}

\item{NA}{\code{T const&} value specifying NA (can be any value).}
}
\value{
a \code{\link[=SPLITT::bvec]{bvec}} of the same length as \code{x}
with \code{false} everywhere, except for the positions where there are \code{NA}s.
}
\description{
This is a template function with a specification for the type \link[=SPLITT::uint]{uint}.

\code{
template<class T> inline bvec IsNA(std::vector<T> const& x, T const& NA);

inline bvec IsNA(uvec const& x);}

The element type \code{T} must define an \code{==} operator.
}
\seealso{
\link{SPLITT}

Other global functions: \code{\link{SPLITT::At}},
  \code{\link{SPLITT::Match}},
  \code{\link{SPLITT::NotIsNA}}, \code{\link{SPLITT::Seq}},
  \code{\link{SPLITT::SortIndices}}
}
