% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::Match}
\alias{SPLITT::Match}
\title{Match the first occurrences of elements in a vector of integers.}
\arguments{
\item{x}{\code{VectorValues const&} of values to search for in \code{table}. 
This vector should not have duplicates or \code{NA}s.}

\item{table}{\code{VectorValues const&} of (possibly duplicated) values to 
matched against the elements in \code{x}.}

\item{NA}{\code{PosType const&} an integer type that can be used as index-type
in \code{x} and \code{table}.}
}
\value{
an \href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<PosType>}.
}
\description{
This is a template function optimized for searching integer types (internal 
use only).

\code{
template<class VectorValues, class PosType>
inline std::vector<PosType> Match(VectorValues const& x, 
VectorValues const& table, PosType const& NA);}

For each element of \code{x} return the index of its first occurence in 
\code{table} or \code{NA} if the element is not found in \code{table} or is 
equal to \code{NA}. It is assumed that \code{x} does not have duplicated 
elements or \code{NA} elements.

The template argument \code{VectorValues} must 
  be an index-able class, such as 
  \href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<T>},
  where \code{T} must be an integer type.
  The template argument \code{PosType} must be an integer type that can be used
  as index type in \code{VectorValues}.
}
\seealso{
\link{SPLITT}

Other global functions: \code{\link{SPLITT::At}},
  \code{\link{SPLITT::IsNA}},
  \code{\link{SPLITT::NotIsNA}}, \code{\link{SPLITT::Seq}},
  \code{\link{SPLITT::SortIndices}}
}
\concept{global functions}
