% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::OrderedTree::num_parallel_ranges_prune}
\alias{SPLITT::OrderedTree::num_parallel_ranges_prune}
\title{Number of parallel ranges of nodeduring}
\description{
\code{
\link[=SPLITT::uint]{uint} num_levels() const;}

During post-order traversal, a node is pruned from its parent after it gets 
visited. This is the so called operation \code{PruneNode}. Prune-ranges 
represent groups of nodes that can be pruned independent from one another 
and, therefore, in parallel. Conceptually prune-ranges are similar to levels 
of (parallel) \code{VisitNode} operations (see also
\link[=SPLITT::OrderedTree::num_levels]{num_levels}()), but 
In a balanced tree the number of levels is in the order of O(log2(N)), where N
is the number of tips in the tree and, therefore, parallelization can be 
beneficial. In a strongly unbalanced tree, the number of levels is in the 
order of O(N), and the parallelization of the \code{VisitNode} operation
cannot improve the speed.
}
\seealso{
\code{\link[=SPLITT::OrderedTree]{OrderedTree}} \code{\link[=SPLITT::Tree]{Tree}} \code{\link[=SPLITT::Tree::Tree]{Tree::Tree()}}

Other \code{\link[=SPLITT::OrderedTree]{OrderedTree}} public methods: \code{\link{SPLITT::OrderedTree::OrderedTree}},
  \code{\link{SPLITT::OrderedTree::num_levels}}
}
