% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::OrderedTree::RangeIdPruneNode}
\alias{SPLITT::OrderedTree::RangeIdPruneNode}
\title{First and last id (0-based node indices) of the PruneNode-range at a step}
\arguments{
\item{i_step}{the step (0-based) of the PruneNode-range}
}
\description{
\href{http://en.cppreference.com/w/cpp/container/array}{\code{std::array}}\code{<\link[=SPLITT::uint]{uint}, 2>}\code{ RangeIdPruneNode(\link[=SPLITT::uint]{uint} i_step) const;}
}
\seealso{
\code{\link[=SPLITT::OrderedTree]{OrderedTree}}

\link{SPLITT}

Other \code{\link[=SPLITT::OrderedTree]{OrderedTree}} public methods: \code{\link{SPLITT::OrderedTree::OrderedTree}},
  \code{\link{SPLITT::OrderedTree::RangeIdVisitNode}},
  \code{\link{SPLITT::OrderedTree::num_levels}},
  \code{\link{SPLITT::OrderedTree::num_parallel_ranges_prune}},
  \code{\link{SPLITT::OrderedTree::ranges_id_prune}},
  \code{\link{SPLITT::OrderedTree::ranges_id_visit}}
}
