% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::OrderedTree::RangeIdVisitNode}
\alias{SPLITT::OrderedTree::RangeIdVisitNode}
\title{First and last id (0-based node indices) of the VisitNode-range at a level}
\arguments{
\item{i_level}{the level (0-based) of the VisitNode-range}
}
\description{
\href{http://en.cppreference.com/w/cpp/container/array}{\code{std::array}}\code{<\link[=SPLITT::uint]{uint}, 2>}\code{ RangeIdVisitNode(\link[=SPLITT::uint]{uint} i_level) const;}
}
\seealso{
\code{\link[=SPLITT::OrderedTree]{OrderedTree}}

\link{SPLITT}

Other \code{\link[=SPLITT::OrderedTree]{OrderedTree}} public methods: \code{\link{SPLITT::OrderedTree::OrderedTree}},
  \code{\link{SPLITT::OrderedTree::RangeIdPruneNode}},
  \code{\link{SPLITT::OrderedTree::num_levels}},
  \code{\link{SPLITT::OrderedTree::num_parallel_ranges_prune}},
  \code{\link{SPLITT::OrderedTree::ranges_id_prune}},
  \code{\link{SPLITT::OrderedTree::ranges_id_visit}}
}
