% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::OrderedTree::ranges_id_visit}
\alias{SPLITT::OrderedTree::ranges_id_visit}
\title{Internally stored vector of start ids for each VisitNode-range}
\description{
\code{
\link[=SPLITT::uvec]{uvec} const& ranges_id_visit() const;}
}
\seealso{
\code{\link[=SPLITT::OrderedTree]{OrderedTree}}

\link{SPLITT}

Other public methods in SPLITT::OrderedTree: \code{\link{SPLITT::OrderedTree::OrderedTree}},
  \code{\link{SPLITT::OrderedTree::RangeIdPruneNode}},
  \code{\link{SPLITT::OrderedTree::RangeIdVisitNode}},
  \code{\link{SPLITT::OrderedTree::num_levels}},
  \code{\link{SPLITT::OrderedTree::num_parallel_ranges_prune}},
  \code{\link{SPLITT::OrderedTree::ranges_id_prune}}
}
\concept{public methods in SPLITT::OrderedTree}
