% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::OrderedTree}
\alias{SPLITT::OrderedTree}
\title{A generic tree class optimized for parallel ordered traversal.}
\description{
This is a template class inheriting from \code{\link[=SPLITT::Tree]{Tree}}.

\code{
template<class Node, class Length>class OrderedTree: public Tree<Node, Length>}
}
\section{Template Arguments}{

\describe{
\item{class Node}{see \code{\link[=SPLITT::OrderedTree::NodeType]{NodeType}}.}
\item{class Length}{see \code{\link[=SPLITT::OrderedTree::LengthType]{LengthType}}.}
}
}

\section{Public Methods}{

\describe{
\item{\code{\link[=SPLITT::OrderedTree::num_levels]{num_levels}}}{}
\item{\code{\link[=SPLITT::OrderedTree::num_parallel_ranges_prune]{num_parallel_ranges_prune}}}{}
\item{\code{\link[=SPLITT::OrderedTree::ranges_id_visit]{ranges_id_visit}}}{}
\item{\code{\link[=SPLITT::OrderedTree::RangeIdVisitNode]{RangeIdVisitNode}}}{}
\item{\code{\link[=SPLITT::OrderedTree::ranges_id_prune]{ranges_id_prune}}}{}
\item{\code{\link[=SPLITT::OrderedTree::RangeIdPruneNode]{RangeIdPruneNode}}}{}
}
}

\seealso{
\code{\link[=SPLITT::Tree]{Tree}}

\link{SPLITT}
}
