% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::Seq}
\alias{SPLITT::Seq}
\title{Generate a sequence of consecutive integers.}
\arguments{
\item{first, last}{\code{T const&} first and last elements in the sequence.}
}
\value{
a \href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<T>} object.
}
\description{
This is a template function.

\code{
template<class T> inline std::vector<T> Seq(T const& first, T const& last);}

Currently the function is implemented as a wrapper for 
\href{https://en.cppreference.com/w/cpp/algorithm/iota}{\code{std::iota}}\code{<T>}.
}
\seealso{
\link{SPLITT}

Other global functions: \code{\link{SPLITT::At}},
  \code{\link{SPLITT::IsNA}}, \code{\link{SPLITT::Match}},
  \code{\link{SPLITT::NotIsNA}},
  \code{\link{SPLITT::SortIndices}}
}
\concept{global functions}
