% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::SortIndices}
\alias{SPLITT::SortIndices}
\title{Indices in a vector ordered in ascending order of the corresponding values.}
\arguments{
\item{v}{a \code{const&} to a \code{VectorClass} object.}
}
\value{
an \href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<\link[=SPLITT::uint]{uint}>}.
}
\description{
\code{template <class VectorClass>
  inline std::vector<uint> SortIndices(VectorClass const& v);} 
  
  This is a template function. The template argument \code{VectorClass} must 
  be an index-able class, such as 
  \href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector<T>}},
  where the type of the elements, \code{T}, must have operator <.
}
\seealso{
\link{SPLITT}

Other global functions: \code{\link{SPLITT::At}},
  \code{\link{SPLITT::IsNA}}, \code{\link{SPLITT::Match}},
  \code{\link{SPLITT::NotIsNA}}, \code{\link{SPLITT::Seq}}
}
\concept{global functions}
