% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::TraversalSpecification}
\alias{SPLITT::TraversalSpecification}
\title{An abstract base class for specifying node traversal operations.}
\description{
This is an abstract class (not intended for instantiation).
The user must provide a TraversalSpecificationImplementation class implementing
this class' methods as described below. It is
recommended to inherit from this class, but it is not obligatory, because 
this is is not checked during compilation. Use the documentation of this 
class as a guide through the steps of writing a tree traversal specification
based on \code{\link{SPLITT}}.
}
\seealso{
\code{\link{SPLITT}}
}
