% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::TraversalTask}
\alias{SPLITT::TraversalTask}
\title{A composite of a tree, a traversal specification and a traversal algorithm.}
\description{
\code{
template<class TraversalSpecification> class TraversalTask;
}
}
\section{Template Arguments}{

\itemize{
\item{class TraversalSpecification}{see \code{\link[=SPLITT::TraversalSpecification]{TraversalSpecification}}.}
}
}

\section{Public Methods}{

\describe{
\item{\link[=SPLITT::TraversalTask::TraversalTaskLightweight]{TraversalTask}}{}
\item{\link[=SPLITT::TraversalTask::TraverseTree]{TraverseTree}}{}
\item{\link[=SPLITT::TraversalTask::tree]{tree}}{}
\item{\link[=SPLITT::TraversalTask::spec]{spec}}{}
\item{\link[=SPLITT::TraversalTask::algorithm]{algorithm}}{}
}
}

\seealso{
\link{SPLITT}

\link{SPLITT::TraversalSpecification}
}
