% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::TraversalTaskLightweight}
\alias{SPLITT::TraversalTaskLightweight}
\title{A lighter TraversalTask class which gets a reference to an already 
constructed tree.}
\description{
\code{
template<class TraversalSpecification> class TraversalTaskLightweight;
}
}
\section{Template Arguments}{

\itemize{
\item{class TraversalSpecification}{see \code{\link[=SPLITT::TraversalSpecification]{TraversalSpecification}}.}
}
}

\section{Public Methods}{

\describe{
\item{\link[=SPLITT::TraversalTaskLightweight::TraversalTaskLightweight]{TraversalTaskLightweight}}{}
\item{\link[=SPLITT::TraversalTaskLightweight::TraverseTree]{TraverseTree}}{}
\item{\link[=SPLITT::TraversalTaskLightweight::spec]{spec}}{}
\item{\link[=SPLITT::TraversalTaskLightweight::algorithm]{algorithm}}{}
}
}

\seealso{
\link{SPLITT}

\link{SPLITT::TraversalSpecification}
}
