% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::Tree::FindChildren}
\alias{SPLITT::Tree::FindChildren}
\title{Get a vector with the ids of the children of a node.}
\arguments{
\item{i}{\code{\link[=SPLITT::uint]{uint}}: the id of a node.}
}
\value{
a \code{uvec const&}: a const reference to an internally stored vector 
of the ids of the children of \code{i}. If \code{i} is a tip, 
  \code{\link[=SPLITT::EMPTY_UVEC]{EMPTY_UVEC}} is returned. The returned 
  vector reference is valid as long as the tree object is not destroyed.
}
\description{
\code{\link[=SPLITT::uvec]{uvec} const& FindChildren(uint i) const;}
}
\seealso{
\code{\link[=SPLITT::Tree]{Tree}}

Other \code{\link[=SPLITT::Tree]{Tree}} public methods: \code{\link{SPLITT::Tree::BranchLengths}},
  \code{\link{SPLITT::Tree::FindIdOfNode}},
  \code{\link{SPLITT::Tree::FindIdOfParent}},
  \code{\link{SPLITT::Tree::FindNodeWithId}},
  \code{\link{SPLITT::Tree::HasBranchLengths}},
  \code{\link{SPLITT::Tree::LengthOfBranch}},
  \code{\link{SPLITT::Tree::OrderNodesPosType}},
  \code{\link{SPLITT::Tree::OrderNodes}},
  \code{\link{SPLITT::Tree::SetBranchLengths}},
  \code{\link{SPLITT::Tree::SetLengthOfBranch}},
  \code{\link{SPLITT::Tree::Tree}},
  \code{\link{SPLITT::Tree::num_nodes}},
  \code{\link{SPLITT::Tree::num_tips}}
}
