% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::Tree::FindNodeWithId}
\alias{SPLITT::Tree::FindNodeWithId}
\title{Get the node with the specified id.}
\arguments{
\item{id}{\code{\link[=SPLITT::uint]{uint}} the id of the node (should be 
between 0 and M-1, where M is the number of nodes in hte tree).}
}
\description{
\code{\link[=SPLITT::Tree::NodeType]{NodeType} const& 
FindNodeWithId(\link[=SPLITT::uint]{uint} id) const;}
}
\seealso{
\code{\link[=SPLITT::Tree]{Tree}}

\link{SPLITT}

Other \code{\link[=SPLITT::Tree]{Tree}} public methods: \code{\link{SPLITT::Tree::BranchLengths}},
  \code{\link{SPLITT::Tree::FindChildren}},
  \code{\link{SPLITT::Tree::FindIdOfNode}},
  \code{\link{SPLITT::Tree::FindIdOfParent}},
  \code{\link{SPLITT::Tree::HasBranchLengths}},
  \code{\link{SPLITT::Tree::LengthOfBranch}},
  \code{\link{SPLITT::Tree::OrderNodesPosType}},
  \code{\link{SPLITT::Tree::OrderNodes}},
  \code{\link{SPLITT::Tree::SetBranchLengths}},
  \code{\link{SPLITT::Tree::SetLengthOfBranch}},
  \code{\link{SPLITT::Tree::Tree}},
  \code{\link{SPLITT::Tree::num_nodes}},
  \code{\link{SPLITT::Tree::num_tips}}
}
\concept{\code{\link[=SPLITT::Tree]{Tree}} public methods}
