% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::Tree::LengthOfBranch}
\alias{SPLITT::Tree::LengthOfBranch}
\title{Get the length of a branch ending at node with id \code{i}.}
\arguments{
\item{i}{\code{\link{uint}}; the id of the end-node for the branch}
}
\value{
\code{\link{LengthType}}; the length associated with the branch ending at node \code{i}.
}
\description{
\code{\link{LengthType} const& \link[=SPLITT::Tree]{Tree}::LengthOfBranch(uint i) const;}
}
\seealso{
\code{\link[=SPLITT::Tree]{Tree}}

Other \code{\link[=SPLITT::Tree]{Tree}} public methods: \code{\link{SPLITT::Tree::BranchLengths}},
  \code{\link{SPLITT::Tree::FindChildren}},
  \code{\link{SPLITT::Tree::FindIdOfNode}},
  \code{\link{SPLITT::Tree::FindIdOfParent}},
  \code{\link{SPLITT::Tree::FindNodeWithId}},
  \code{\link{SPLITT::Tree::HasBranchLengths}},
  \code{\link{SPLITT::Tree::OrderNodesPosType}},
  \code{\link{SPLITT::Tree::OrderNodes}},
  \code{\link{SPLITT::Tree::SetBranchLengths}},
  \code{\link{SPLITT::Tree::SetLengthOfBranch}},
  \code{\link{SPLITT::Tree::Tree}},
  \code{\link{SPLITT::Tree::num_nodes}},
  \code{\link{SPLITT::Tree::num_tips}}
}
