% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::Tree::num_nodes}
\alias{SPLITT::Tree::num_nodes}
\title{Number of nodes in a tree.}
\value{
a \code{\link[=SPLITT::uint]{uint}}: the numbef of nodes in the tree, 
including tips, internal nodes and the root.
}
\description{
\code{\link[=SPLITT::uint]{uint} Tree::num_nodes() const;}
}
\seealso{
\code{\link[=SPLITT::Tree]{Tree}}

\link{SPLITT}

Other public methods in SPLITT::Tree: \code{\link{SPLITT::Tree::BranchLengths}},
  \code{\link{SPLITT::Tree::FindChildren}},
  \code{\link{SPLITT::Tree::FindIdOfNode}},
  \code{\link{SPLITT::Tree::FindIdOfParent}},
  \code{\link{SPLITT::Tree::FindNodeWithId}},
  \code{\link{SPLITT::Tree::HasBranchLengths}},
  \code{\link{SPLITT::Tree::LengthOfBranch}},
  \code{\link{SPLITT::Tree::OrderNodesPosType}},
  \code{\link{SPLITT::Tree::OrderNodes}},
  \code{\link{SPLITT::Tree::SetBranchLengths}},
  \code{\link{SPLITT::Tree::SetLengthOfBranch}},
  \code{\link{SPLITT::Tree::Tree}},
  \code{\link{SPLITT::Tree::num_tips}}
}
\concept{public methods in SPLITT::Tree}
