% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::Tree::SetLengthOfBranch}
\alias{SPLITT::Tree::SetLengthOfBranch}
\title{Set the length of a branch ending at node with id \code{i} to a given \code{value}.}
\arguments{
\item{i}{\code{\link[=SPLITT::uint]{uint}}; the id of the end-node of the branch;}

\item{value}{\code{\link[=SPLITT::Tree::LengthType]{LengthType} const&}; the new value to set.}
}
\value{
\code{void}
}
\description{
\code{void SetLengthOfBranch(\link[=SPLITT::uint]{uint} i, \link[=SPLITT::Tree::LengthType]{LengthType} const& value);}
}
\seealso{
\code{\link[=SPLITT::Tree]{Tree}}

\link{SPLITT}

Other \code{\link[=SPLITT::Tree]{Tree}} public methods: \code{\link{SPLITT::Tree::BranchLengths}},
  \code{\link{SPLITT::Tree::FindChildren}},
  \code{\link{SPLITT::Tree::FindIdOfNode}},
  \code{\link{SPLITT::Tree::FindIdOfParent}},
  \code{\link{SPLITT::Tree::FindNodeWithId}},
  \code{\link{SPLITT::Tree::HasBranchLengths}},
  \code{\link{SPLITT::Tree::LengthOfBranch}},
  \code{\link{SPLITT::Tree::OrderNodesPosType}},
  \code{\link{SPLITT::Tree::OrderNodes}},
  \code{\link{SPLITT::Tree::SetBranchLengths}},
  \code{\link{SPLITT::Tree::Tree}},
  \code{\link{SPLITT::Tree::num_nodes}},
  \code{\link{SPLITT::Tree::num_tips}}
}
\concept{\code{\link[=SPLITT::Tree]{Tree}} public methods}
