% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::Tree::Tree}
\alias{SPLITT::Tree::Tree}
\title{Constructor for class \code{\link[=SPLITT::Tree]{Tree}}.}
\arguments{
\item{branch_start_nodes}{\href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<\link[=SPLITT::Tree::NodeType]{NodeType}> const&}: 
starting node for every branch in the tree.}

\item{branch_end_nodes}{\href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<\link[=SPLITT::Tree::NodeType]{NodeType}> const&}:
ending node for every branch in the tree; must be the same length as 
\code{branch_start_nodes}.}

\item{branch_lengths}{\href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<\link[=SPLITT::Tree::LengthType]{LengthType}> const&}: 
lengths associated with the branches. Pass an empty vector for \code{branch_lengths} 
for a tree without branch lengths (i.e. only a topology).}
}
\description{
\code{
Tree(std::vector<NodeType> const& branch_start_nodes,
std::vector<NodeType> const& branch_end_nodes,
std::vector<LengthType> const& branch_lengths);}

Constructs the tree object given a list of branches. The list of branches
is specified from the corresponding elements in the three vectors passed as
arguments.
}
\seealso{
\code{\link[=SPLITT::Tree]{Tree}}

\link{SPLITT}

Other \code{\link[=SPLITT::Tree]{Tree}} public methods: \code{\link{SPLITT::Tree::BranchLengths}},
  \code{\link{SPLITT::Tree::FindChildren}},
  \code{\link{SPLITT::Tree::FindIdOfNode}},
  \code{\link{SPLITT::Tree::FindIdOfParent}},
  \code{\link{SPLITT::Tree::FindNodeWithId}},
  \code{\link{SPLITT::Tree::HasBranchLengths}},
  \code{\link{SPLITT::Tree::LengthOfBranch}},
  \code{\link{SPLITT::Tree::OrderNodesPosType}},
  \code{\link{SPLITT::Tree::OrderNodes}},
  \code{\link{SPLITT::Tree::SetBranchLengths}},
  \code{\link{SPLITT::Tree::SetLengthOfBranch}},
  \code{\link{SPLITT::Tree::num_nodes}},
  \code{\link{SPLITT::Tree::num_tips}}
}
\concept{\code{\link[=SPLITT::Tree]{Tree}} public methods}
