% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT::Tree}
\alias{SPLITT::Tree}
\title{Base class \code{Tree}.}
\description{
A generic C++ template class defining the data structure and 
basic operations with a tree. 
  
\code{template<class Node, class Length>class Tree;}
}
\section{Template Arguments}{

\itemize{
\item{class Node}{see \code{\link[=SPLITT::Tree::NodeType]{NodeType}}.}
\item{class Length}{see \code{\link[=SPLITT::Tree::LengthType]{LengthType}}.}
}
}

\section{Public Methods}{

\describe{
\item{\link[=SPLITT::Tree::Tree]{Tree}}{}
\item{\link[=SPLITT::Tree::num_tips]{num_tips}}{}
\item{\link[=SPLITT::Tree::num_nodes]{num_nodes}}{}
\item{\link[=SPLITT::Tree::BranchLengths]{BranchLengths}}{}
\item{\link[=SPLITT::Tree::FindChildren]{FindChildren}}{}
\item{\link[=SPLITT::Tree::FindIdOfNode]{FindIdOfNode}}{}
\item{\link[=SPLITT::Tree::FindIdOfParent]{FindIdOfParent}}{}
\item{\link[=SPLITT::Tree::FindNodeWithId]{FindNodeWithId}}{}
\item{\link[=SPLITT::Tree::HasBranchLengths]{HasBranchLengths}}{}
\item{\link[=SPLITT::Tree::LengthOfBranch]{LengthOfBranch}}{}
\item{\link[=SPLITT::Tree::SetBranchLengths]{SetBranchLengths}}{}
\item{\link[=SPLITT::Tree::SetLengthOfBranch]{SetLengthOfBranch}}{}
\item{\link[=SPLITT::Tree::OrderNodesPosType]{OrderNodesPosType}}{}
\item{\link[=SPLITT::Tree::OrderNodes]{OrderNodes}}{}
}
}

\seealso{
\code{\link[=SPLITT::OrderedTree]{OrderedTree}}

\link{SPLITT}
}
