% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::uint}
\alias{SPLITT::uint}
\title{An alias for the \code{'unsigned int'} basic type.}
\description{
\code{typedef unsigned int uint;}
}
\seealso{
Other basic types: \code{\link{SPLITT::bvec}},
  \code{\link{SPLITT::uvec}}, \code{\link{SPLITT::vec}}
}
\concept{basic types}
