% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::uvec}
\alias{SPLITT::uvec}
\title{a vector of \code{\link[=SPLITT::uint]{uint}}'s.}
\description{
\code{typedef }\href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<uint> uvec;}
}
\seealso{
Other basic types: \code{\link{SPLITT::bvec}},
  \code{\link{SPLITT::uint}}, \code{\link{SPLITT::vec}}
}
