% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::vec}
\alias{SPLITT::vec}
\title{a vector of \code{double}s.}
\description{
\code{typedef }\href{http://en.cppreference.com/w/cpp/container/vector}{\code{std::vector}}\code{<double> vec;}
}
\seealso{
\link{SPLITT}

Other basic types: \code{\link{SPLITT::bvec}},
  \code{\link{SPLITT::uint}}, \code{\link{SPLITT::uvec}}
}
\concept{basic types}
