% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPLITT}
\alias{SPLITT}
\title{SPLITT: A generic C++ library for Serial and Parallel Lineage Traversal of Trees}
\description{
SPLITT defines a set of basic types, classes, global 
functions and constants:
}
\details{
This documentation of the SPLITT API is still under construction. 
Not all classes and methods have yet been described.
}
\section{Basic types}{

\describe{
\item{\link[=SPLITT::uint]{uint}}{}
\item{\link[=SPLITT::uvec]{uvec}}{}
\item{\link[=SPLITT::vec]{vec}}{}
\item{\link[=SPLITT::bvec]{bvec}}{}
}
}

\section{Global constants}{

\describe{
\item{\link[=SPLITT::G_NA_UINT]{G_NA_UINT}}{}
\item{\link[=SPLITT::G_EMPTY_UVEC]{G_EMPTY_UVEC}}{}
\item{\link[=SPLITT::G_PI]{G_PI}}{}
}
}

\section{Global functions}{

\describe{
\item{\link[=SPLITT::SortIndices]{SortIndices}}{}
\item{\link[=SPLITT::At]{At}}{}
\item{\link[=SPLITT::Match]{Match}}{}
\item{\link[=SPLITT::Seq]{Seq}}{}
\item{\link[=SPLITT::IsNA]{IsNA}}{}
\item{\link[=SPLITT::NotIsNA]{NotIsNA}}{}
}
}

\section{Classes}{

\describe{
\item{\link[=SPLITT::TraversalSpecification]{TraversalSpecification}}{}
\item{\link[=SPLITT::TraversalTask]{TraversalTask}}{}
\item{\link[=SPLITT::TraversalTaskLightweight]{TraversalTaskLightweight}}{}
\item{\link[=SPLITT::Tree]{Tree}}{}
\item{\link[=SPLITT::OrderedTree]{OrderedTree}}{}
\item{\link[=SPLITT::ThreadExceptionHandler]{ThreadExceptionHandler}}{}
}

[[Rcpp::export]]
}

