% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::BranchLengths}
\alias{Tree::BranchLengths}
\alias{BranchLengths}
\title{Get a reference to the internal vector of branch lengths.}
\value{
\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{LengthType}> const&}; 
  a const reference to the internally stored vector of branch lengths, in the order of the end-node ids.
}
\section{Signature}{

\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{LengthType}> const& \link{Tree}::BranchLengths() const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other branch-length-methods: \code{\link{Tree::HasBranchLengths}},
  \code{\link{Tree::LengthOfBranch}},
  \code{\link{Tree::SetBranchLengths2}},
  \code{\link{Tree::SetBranchLengths}},
  \code{\link{Tree::SetLengthOfBranch}}
}
