% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::FindChildren}
\alias{Tree::FindChildren}
\alias{FindChildren}
\title{Get a vector with the ids of the children of a node.}
\arguments{
\item{i}{\code{\link{uint}}; the id of a node.}
}
\value{
a \code{uvec const&}; a const reference to an internally stored vector of the
  ids of the children of \code{i}. If \code{i} is a tip, an empty uvec is returned. The
  returned vector reference is valid as long as the tree object is not destroyed.
}
\section{Signature}{

\code{uvec const& \link{Tree}::FindChildren(uint i) const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other node-methods: \code{\link{Tree::FindIdOfNode}},
  \code{\link{Tree::FindIdOfParent}},
  \code{\link{Tree::FindNodeWithId}},
  \code{\link{Tree::OrderNodesPosType}},
  \code{\link{Tree::OrderNodes}}
}
