% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::FindIdOfNode}
\alias{Tree::FindIdOfNode}
\alias{FindIdOfNode}
\title{Get the internally stored id of a node}
\arguments{
\item{node}{\code{\link{NodeType} const&}; the node;}
}
\value{
an \code{\link{uint}} from 0 to M-1, where M is the number of nodes 
  in the tree.
}
\description{
A public method of class \code{\link{Tree}}.
}
\section{Signature}{

\code{uint \link{Tree}::FindIdOfNode(\link{NodeType} const& node) const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other node-methods: \code{\link{Tree::FindChildren}},
  \code{\link{Tree::FindIdOfParent}},
  \code{\link{Tree::FindNodeWithId}},
  \code{\link{Tree::OrderNodesPosType}},
  \code{\link{Tree::OrderNodes}}
}
