% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::FindIdOfParent}
\alias{Tree::FindIdOfParent}
\alias{FindIdOfParent}
\title{Get the parent id of a node with id id_child.}
\arguments{
\item{id_child}{\code{\link{uint}}, the id of the child node.}
}
\value{
an \code{\link{uint}} from 0 to M-1, where M is the number of nodes 
  in the tree.
}
\section{Signature}{

\code{uint \link{Tree}::FindIdOfParent(uint id_child) const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other node-methods: \code{\link{Tree::FindChildren}},
  \code{\link{Tree::FindIdOfNode}},
  \code{\link{Tree::FindNodeWithId}},
  \code{\link{Tree::OrderNodesPosType}},
  \code{\link{Tree::OrderNodes}}
}
