% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::FindNodeWithId}
\alias{Tree::FindNodeWithId}
\alias{FindNodeWithId}
\title{Get the node with the specified id.}
\arguments{
\item{id}{\code{\link{uint}} the id of the node (should be between 0 and M-1, where
M is the number of nodes in hte tree).}
}
\description{
A public method of class \code{\link{Tree}}.
}
\section{Signature}{

\code{\link{NodeType} const& \link{Tree}::FindNodeWithId(uint id) const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other node-methods: \code{\link{Tree::FindChildren}},
  \code{\link{Tree::FindIdOfNode}},
  \code{\link{Tree::FindIdOfParent}},
  \code{\link{Tree::OrderNodesPosType}},
  \code{\link{Tree::OrderNodes}}
}
