% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::HasBranchLengths}
\alias{Tree::HasBranchLengths}
\alias{HasBranchLengths}
\title{Does a tree has lengths associated with its branches?}
\value{
\code{bool}, \code{true} if the tree has branch lengths.
}
\section{Signature}{

\code{bool \link{Tree}::HasBranchLengths() const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other branch-length-methods: \code{\link{Tree::BranchLengths}},
  \code{\link{Tree::LengthOfBranch}},
  \code{\link{Tree::SetBranchLengths2}},
  \code{\link{Tree::SetBranchLengths}},
  \code{\link{Tree::SetLengthOfBranch}}
}
