% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::LengthOfBranch}
\alias{Tree::LengthOfBranch}
\alias{LengthOfBranch}
\title{Get the length of a branch ending at node with id \code{i}}
\arguments{
\item{i}{\code{\link{uint}}; the id of the end-node for the branch}
}
\value{
\code{\link{LengthType}}; the length associated with the branch ending at node \code{i}.
}
\section{Signature}{

\code{\link{LengthType} const& \link{Tree}::LengthOfBranch(uint i) const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other branch-length-methods: \code{\link{Tree::BranchLengths}},
  \code{\link{Tree::HasBranchLengths}},
  \code{\link{Tree::SetBranchLengths2}},
  \code{\link{Tree::SetBranchLengths}},
  \code{\link{Tree::SetLengthOfBranch}}
}
