% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::LengthType}
\alias{Tree::LengthType}
\alias{LengthType}
\alias{Tree.LengthType}
\title{typedef Length LengthType;}
\description{
A public typedef in class \code{\link{Tree}}. A synonym for the template
  argument Length. Defines a type that can be associated with a branch. Can
  be \code{double}, but also a composite of several attributes on a
  branch, such as a \code{double} length and an \code{int} color.
}
\seealso{
\code{\link{Tree}}
}
