% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::NodeType}
\alias{Tree::NodeType}
\alias{NodeType}
\title{typedef Node NodeType;}
\description{
A public typedef in class \code{\link{Tree}}. A synonym for the 
  template argument \code{Node}. Defines a 
  hash-able type such as \code{int} or 
  \code{\href{http://en.cppreference.com/w/cpp/string/basic_string}{std::string}}. 
  Specifically, this should be a type for which 
  \code{\href{http://en.cppreference.com/w/cpp/utility/hash}{std::hash}}
  specialization does exist. This is the application-specific node-type. The branches in the tree are defined as couples of nodes 
  (branch_start_node, branch_end_node).
}
\details{
During the construction of \code{Tree} object, the nodes are assigned 
  \code{unsigned int} ids from 0 to M-1 (M being the number of nodes in the
  tree).
}
\seealso{
\code{\link{Tree}} \code{\link{FindIdOfNode}}
}
