% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::num_nodes}
\alias{Tree::num_nodes}
\title{Number of nodes in a tree}
\value{
the numbef of nodes in the tree, including tips, internal nodes and the root.
}
\section{Signature}{

\code{\link{uint} Tree::num_nodes() const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}
}
