% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::num_tips}
\alias{Tree::num_tips}
\title{Number of tips (a.k.a. leaves) in the tree}
\value{
\code{\link{uint}}, the number of tips in the tree.
}
\section{Signature}{

\code{\link{uint} \link{Tree}::num_nodes() const;}
}

\seealso{
\code{\link{Tree}}
}
