% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::OrderNodes}
\alias{Tree::OrderNodes}
\alias{OrderNodes}
\title{Reorder a vector of nodes}
\arguments{
\item{nodes}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{NodeType}> const&}; 
a vector of application-specific nodes.}
}
\value{
\code{\link{uvec}}; a vector of positions in \code{nodes} in the order of 
  their internally stored ids.
}
\section{Signature}{

\code{\link{uvec} \link{Tree}::OrderNodes(\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{NodeType}> const& nodes) const;}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other node-methods: \code{\link{Tree::FindChildren}},
  \code{\link{Tree::FindIdOfNode}},
  \code{\link{Tree::FindIdOfParent}},
  \code{\link{Tree::FindNodeWithId}},
  \code{\link{Tree::OrderNodesPosType}}
}
