% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::OrderNodesPosType}
\alias{Tree::OrderNodesPosType}
\alias{OrderNodesPosType}
\title{Reorder a vector of nodes (generic w.r.t. the position type).}
\arguments{
\item{nodes}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<PosType> \link{Tree}::OrderNodesPosType(\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{NodeType}> const&}; 
a vector of application-specific nodes.}

\item{NA}{\code{PosType const&}; NA value used mainly for the purpose of template-specification.}
}
\value{
\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<PosType>}; a vector of positions in nodes in the order of their internally stored ids.
the element-type of the returned vector can be specified as a template argument or 
dedcued from the type of NA.
}
\section{Template Arguments}{

\describe{
\item{PosType}{an integer type for the positions in the returned vector.}}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other node-methods: \code{\link{Tree::FindChildren}},
  \code{\link{Tree::FindIdOfNode}},
  \code{\link{Tree::FindIdOfParent}},
  \code{\link{Tree::FindNodeWithId}},
  \code{\link{Tree::OrderNodes}}
}
