% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::SetBranchLengths}
\alias{Tree::SetBranchLengths}
\alias{SetBranchLengths}
\title{Set a new internally stored vector of branch lenthts}
\arguments{
\item{lengths}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{LengthType}> const&}; 
a const reference to a new vector of branch lengths, in the order of the end-node ids. 
The vector should be of length M-1, where M is the 
number of nodes in the tree.}
}
\value{
\code{void}
}
\section{Signature}{

\code{void \link{Tree}::SetBranchLengths(\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{LengthType}> const& lengths);}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other branch-length-methods: \code{\link{Tree::BranchLengths}},
  \code{\link{Tree::HasBranchLengths}},
  \code{\link{Tree::LengthOfBranch}},
  \code{\link{Tree::SetBranchLengths2}},
  \code{\link{Tree::SetLengthOfBranch}}
}
