% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::SetBranchLengths2}
\alias{Tree::SetBranchLengths2}
\alias{SetBranchLengths}
\title{Set new branch lenthts to the branches in the order given by their 
  application-specific end-nodes.}
\arguments{
\item{nodes_branch_ends}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{NodeType}> const&}; 
a const reference to a new vector of branch lengths, in the order of the nodes in \code{nodes_branch_ends}.}

\item{lengths}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{LengthType}> const&}; 
a const reference to a new vector of branch lengths, in the order of the nodes in \code{nodes_branch_ends}.}
}
\value{
\code{void}
}
\description{
If the tree has no branch lengths, the supplied arguments should
  be of length M-1, where M is the total number of nodes in the tree (-1, 
  because there is no branch leading to the root).
}
\section{Signature}{

\code{void \link{Tree}::SetBranchLengths(\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{NodeType}> const& nodes_branch_ends, \href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{LengthType}> const& lengths);}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other branch-length-methods: \code{\link{Tree::BranchLengths}},
  \code{\link{Tree::HasBranchLengths}},
  \code{\link{Tree::LengthOfBranch}},
  \code{\link{Tree::SetBranchLengths}},
  \code{\link{Tree::SetLengthOfBranch}}
}
