% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::SetLengthOfBranch}
\alias{Tree::SetLengthOfBranch}
\alias{SetLengthOfBranch}
\title{Set the length of a branch ending at node with id \code{i} to a given \code{value}}
\arguments{
\item{i}{\code{\link{uint}}; the id of the end-node of the branch;}

\item{value}{\code{\link{LengthType} const&}; the new value of to set.}
}
\value{
\code{void}
}
\section{Signature}{

\code{void \link{Tree}::SetLengthOfBranch(uint i, \link{LengthType} const& value);}
}

\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}

Other branch-length-methods: \code{\link{Tree::BranchLengths}},
  \code{\link{Tree::HasBranchLengths}},
  \code{\link{Tree::LengthOfBranch}},
  \code{\link{Tree::SetBranchLengths2}},
  \code{\link{Tree::SetBranchLengths}}
}
