% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree::Tree}
\alias{Tree::Tree}
\title{Constructor for class Tree}
\arguments{
\item{branch_start_nodes}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{NodeType}> const&}; 
starting node for every branch in the tree.}

\item{branch_end_nodes}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{NodeType}> const&}; ending 
node for every branch in the tree; must be the same length as 
branch_start_nodes.}

\item{branch_lengths}{\code{\href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{LengthType}> const&}; lengths 
associated with the branches. Pass an empty vector for branch_lengths for 
a tree without branch lengths (i.e. only a topology).}
}
\seealso{
\code{\link{Tree}} \code{\link{OrderedTree}}
}
