% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tree}
\alias{Tree}
\title{template<class Node, class Length>class Tree}
\description{
A generic C++ class defining the data structure and basic 
  operations with a tree.
}
\section{Template Arguments}{

\describe{
\item{class Node}{see \code{\link{NodeType}}.}
\item{class Length}{see \code{\link{LengthType}}.}
}
}

\seealso{
\code{\link{OrderedTree}}
}
