% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uvec}
\alias{uvec}
\title{typedef \href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<\link{uint}> uvec;}
\description{
a vectof of \link{uint}'s.
}
\seealso{
Other basic-types: \code{\link{bvec}}, \code{\link{uint}},
  \code{\link{vec}}
}
