% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vec}
\alias{vec}
\title{typedef \href{http://en.cppreference.com/w/cpp/container/vector}{std::vector}<double> vec;}
\description{
a vectof of double's.
}
\seealso{
Other basic-types: \code{\link{bvec}}, \code{\link{uint}},
  \code{\link{uvec}}
}
