% Generated by roxygen2: do not edit by hand
% Please edit documentation in src/SPLITT.h
\name{SPLITT::TraversalSpecification}
\alias{SPLITT::TraversalSpecification}
\title{An abstract base class for specifying node traversal operations.}
\description{
This is an abstract class (not intended for instantiation) that defines the 

entities 
The user must provide a TraversalSpecificationImplementation class implementing
this class' methods as described in the below. It is
recommended to inherit from this class, but it is not obligatory (it is not 
checked during compilation).
}
\seealso{
\code{\link{SPLITT}}
}
